Calculate the available buffer space
Determine the amount of available virtual memory and report that as the
available buffer space instead of always returning a constant value.

If the amount of available virtual memory exceeds the range of the
return type the size will be reported as 1 gigabyte to avoid any
potential plug-in issues with handling the largest positive value
of the return type.