Add support for using the Paint.NET color themes
Added a TabControlEx class to support rendering the tab control with a
custom background color.
Changed the HeaderLabel class to use the ForeColor as the text color
when the control is not using the default BackColor.
Fixed a visual glitch when the ColorWheel background color is changed.

Updated the PluginUISettings and ColorPickerService classes to support
setting the color picker background and foreground colors.
Updated the PsFilterPdnConfigDialog and ColorPickerForm classes to
support changing the background and foreground colors of the controls.