Remove the finalizer from the HandleEntry class
When the SetHandleSize method is used by a filter the finalizer will
run on the HandleEntry that was replaced.
This causes a double-free bug because the new and old HandleEntry
instances store some of the same memory addresses and an exception will
be thrown when HeapFree tries to free the same memory twice.