Revert "Do not save filter parameters that are empty"
The potential minor memory savings from not storing "empty" parameters
in the effect token when a filter does not save any parameter data is
not worth the additional complexity in determining whether a filter has
parameter data that needs to be saved.

A filter not having any saved parameter data is a somewhat rare case it
usually occurs when the filter does not have a user interface.

This reverts commit af3c1329bf2d67644b0841e700a2c79393171411.