Move the PICA suites into the PICA folder
Moved the classes, delegates and structures for the Action Descriptor,
Action List and Action Reference suites into an Action Suites
sub-directory in the PICA folder.

Moved the ActionSuiteProvider class and PICASuiteStructs.cs into the
PICA folder.
Moved the classes, delegates and structures for the Descriptor Registry
and Error suites into the PICA folder.

Updated the files to use the PSFilterLoad.PSApi.PICA namespace.