Change the inactive tab background color
The inactive tab background color will be be made darker or lighter
than the active tab background color depending on the user's selected
theme.

If the user has the Dark theme selected the inactive tab color will be
darkened, otherwise it will be lightened.
This was done to prevent the inactive tab background color from being
gray, which tends to occur when you darken a light color or lighten a
dark color.

This helps to differentiate the active and inactive tabs.