Hide the Dfine 2 scripting-only filters
The Dfine 2 filter in the Google Nik Collection uses the **Hidden*
category for its additional helper filters.
The category name is missing the second * at the end when compared to
the normal category name for hidden filters (**Hidden**).

The **Hidden** category name is defined in PIPL.r in the Photoshop SDK.