Use the OS dialog for the color picker prompts
This replaces the custom WinForms dialog that the plugin was previously
using.
The custom dialog was based on the colors floating tool window from
Paint.NET 3.36, and the color wheel appearance does not match the newer
versions of Paint.NET.

Most importantly, this change removes the WinForms dependency for the
PSFilterShim executable.
Removing the WinForms dependency allows trimming to be used and reduces
the single-file executable size from ~150 MB with WinForms down to
~60 MB without trimming.

Trimming is currently disabled due to it not working with the classes
that still use ISerializable.