Add support for writing outside of the selection
Paint.NET 5.0 added a DisableSelectionClipping render flag which allows
BitmapEffect and GpuEffect plugins to write to the entire destination
canvas.

This allows the 8bf filters that render outside of the selection to
work correctly.
Note that some filters may ignore the selection entirely (e.g. Luce 2).

The filters that do not request to render outside of the selection will
have the output image clipped to the selection mask in post-processing.