Allow filters to read the ICC profile
If the image has an embedded ICC profile it will be used as the
document color profile.
For images without an embedded ICC profile we look for the EXIF color
space marker that indicates the use of sRGB or AdobeRGB, and load a
compatible profile from the embedded resources.

If the image does not have an embedded ICC profile or EXIF color space
marker it is assumed to be a sRGB image.