Improve the AdobeRGB profile detection
According to saucecontrol on the C# Discord the AdobeRgb EXIF color
space value that WIC provides is non-standard and not usable for most
images.
The standard way to indicate AdobeRGB is by setting the EXIF color
space to Uncalibrated and the EXIF interoperability index to R03.

See the following links for more information:

https://discord.com/channels/143867839282020352/960223751599976479/1035955858980474902
https://ninedegreesbelow.com/photography/embedded-color-space-information.html