Reduce the PSFilterShimImage I/O memory usage
In most cases we do not need to copy the image data to an intermediate
buffer when reading and writing, the Stream can use a Span/ReadOnlySpan
to access the image data that resides in native memory.
For the cases where an intermediate buffer is still required we use
ArrayPool to reduce the GC overhead.