Fix a few issues with floating selection handling
Floating selections are used to fake the display of transparency
information for the filters that do not natively support it.
Any transparent pixel that the filter modifies should be set to its
original value before the image is clipped to the actual selection.

The FilterRecord autoMask field has been set to true for floating
selections, matching the documented behavior in the SDK.