Fix a Paint.NET hang if PSFilterShim crashes
If PSFilterShim crashed when the plugin dialog was open Paint.NET would
behave as if it had an invisible modal dialog active.
This was caused by PSFilterShim using the Paint.NET window as the
parent for the plugin dialog that it launched.

The PSFilterShim process now creates its own window that it uses as the
parent for the plugin dialog.