Set the PSFilterShim OS subsystem version to 5.1
This stops the OS from automatically enabling termination on heap
corruption, which is incompatible with some older filters.
This fixes a regression from version 1.0.8.2.

Most of the filters that have this issue are built with old versions of
FilterMeister.
The FM Patcher utility can be used to fix the heap corruption bug in
these filters, which may be required for other 32-bit plugin hosts.