Fix a crash when serializing an action descriptor object
DataContractSerializer was failing to serialize an AETEValue object
with a Dictionary<uint, AETEValue> item.
I was unable to fix this issue with the DataContractSerializer, so I
converted all of the serialization code for the scripting objects to
use MessagePack.

This allows me to have full control over how each custom type is
serialized, and as a bonus the output is much smaller than the XML
produced by DataContractSerializer.