Share ChannelPortsSuite.cs and LoadFilter.cs between the projects
This required adding a few types that act as a common interface for the
image and selection mask data.
PSFilterPdn's implementation uses a WIC bitmap, and the PSFilterShim
implementation uses types based on the Paint.NET 3.36 Surface class.

The image used in DisplayPixelsProc now has a 24-bits-per-pixel
version, this saves memory when the image to display does not have an
alpha channel.
LoadPsFilter now ignores the FilterCaseInfo DontCopyToDestination flag,
the source image is always copied to the destination.