Add surface classes that wrap IEffectInputBitmap<T>
This reduces memory usage for the 64-bit filters by eliminating a few
cases of bitmap copying.

Because the IEffectInputBitmap<T> is owned by Paint.NET we do not need
to make a copy of it when cloning the image.
It also may eliminate a copy when locking the image.

This change also allows the value that specifies if the image contains
transparency to be cached.

Because the source and destination images now use different
implementations, the previous behavior of optionally copying the source
to the destination has been restored.