Enable trimming for PSFilterShim
Trimming is performed using the partial trim mode, in this mode ILLink
will only trim the libraries that are marked as compatible with it.
The PSFilterShim assembly is excluded from trimming using the
TrimmerRootAssembly item.

Currently, the MessagePack libraries have not been updated with
trimming support, ILLink will copy these files to the output
unmodified.