Replace the OS P/Invokes with TerraFX.Interop.Windows
This removed a lot of duplicated code, and allows PluginModule.cs to be
shared between the projects.

PSFilterPdn is now deployed to a sub-folder of the Paint.NET Effects
folder with its dependencies.
The dependencies are no longer merged into PSFilterPdn.dll because
merging the DLLs with ILRepack breaks the debugging support, and
ILMerge.Fody does not support TerraFX.Interop.Windows.

Moving PSFilterPdn into its own folder also prevents any of the native
DLLs that PSFilterShim may require from conflicting with the native
DLLs used by other Paint.NET plugins.

CommunityToolkit.HighPerformance and TerraFX.Interop.Windows will be
trimmed with ILLink for the Debug and Release configurations, the
trimming step is skipped for the new FastDebug configuration.