Use Direct2D for the DisplayPixelsProc HDC rendering
This should be faster than the previous GDI+ code, and it also reduces
memory usage by not requiring manual double-buffering or a full-size
transparency checkerboard image.

The Direct2D rendering is preformed using an abstraction that uses the
Paint.NET Direct2D APIs in PSFilterPdn, and TerraFX.Interop.Windows in
PSFilterShim.