Add a custom application icon for PSFilterShim
The custom icon is the same as the Effect icon used by PSFilterPdn.
This icon is stored as part of the native Win32 resources, and will be
used in Windows Explorer and the task bar.

When the application window is created an appropriately sized icon will
be loaded from the native resources.