Allow plugins to read the IPTC caption record
The legacy IPTC IIM format has mostly been superseded by XMP, but some
plugins and/or images may still use this older format.

The IPTC data will only be present for files that already contain it in
the legacy IPTC IIM format, we currently do not check other metadata
for the image caption information.
For example, XMP uses the dc:description tag as its equivalent for
storing the image caption.