Simplify the ExifWriter initialization
ExifWriter now performs the tag filter itself, instead of relying on
the caller to do it before creating an ExifWriterInfo instance.

The image width and height tags are no longer written explicitly.
Based on the behavior of exiftool when adding EXIF data to a PNG image,
the width and height tags are not mandatory.
These tags will only be included if they are present in the Paint.NET
EXIF data.

Moved the image section tag filtering code into the ExifWriter class.