Don't ignore shortcuts to non-existent 8bf files
The code was calling GetFileAttributesW to check if the shortcut
pointed to a directory, this had the side effect of silently ignoring
8bf files that do not exist on disk.

We now get the file attributes from the shortcut file, which allows the
plugin to report any path errors when loading the 8bf files.
Shortcuts to directories that do not exist will be silently ignored,
which is the behavior the plugin has for any error that occurs when
scanning directories.