Fix a few breaks in the ILLink post-build steps
The issues were caused by having a newer .NET SDK installed than the
runtime we are targeting.

Unfortunately, it appears that the only simple way to fix the SDK
version is to hard-code the version number of the .NET 7.0 SDK that we
are using.

We now get the illink assembly from a NuGet package, this removes the
dependency in the runtime having it as part of the default SDK install.