Fix a crash with the 64-bit Mehdi plugins
These filters crash when reshowing the UI after the saved 'global'
parameters have been restored.
They appear to have been built with FilterMeister version 1.09g, but
I don't think it is a FilterMeister bug.
It is likely due to a behavior change in .NET 5+ or the OS.
The process exit code appears to be a Windows 'Exception thrown from
user callback' error.

Only the 64-bit versions of the plugins crash, the 32-bit versions work
fine.
The non-interactive 'repeat effect' command also works without
crashing.

To fix this, we do not restore the 'global' parameters when reshowing
the UI.
The plugins will use their AETE scripting parameters instead.